rename wheel.js to sf-wheel.js
